Move `ExternalPoolData` to `types` dir
